Correção exercício 3

+++ Exercícios 1/Exercicio3.java
@@ -2,15 +2,15 @@ import java.util.Scanner;
 
 public class Exercicio3 {
     public static void main(String[] args) {
-        float media = 0;
+        double media = 0;
 
         for(int i = 0; i < 10; i++) {
 
             Scanner input = new Scanner(System.in);
-            media += input.nextFloat();
+            media += input.nextDouble();
         }
 
-        System.out.printf("Media: %.1f\n", media / 10);
+        System.out.printf("Media: %.1lf\n", media / 10);
     }
 }
